api_service/requirements.txt
fastapi>=0.95
uvicorn>=0.22
pydantic
psycopg2-binary
sqlalchemy
redis
websockets
python-dotenv